refactor(export): split letterhead name-guard family out of letter.rs (R8)
Pre-push architecture gate (R8, hard cap 1400 LOC): letter.rs had grown to
1497 LOC from the name-guard implementation + test additions in the last two
commits. Per the recorded rule, extracted a cohesive block VERBATIM into a
new sibling file rather than trimming anything.

New apps/desktop/src-tauri/src/export/typst_engine/letterhead.rs holds the
letterhead name-guard family — looks_like_date, monogram_initials,
is_letterhead_name, letterhead_initials — plus their 18 directly-related
unit tests (moved byte-for-byte; behaviour and doc comments unchanged except
looks_like_date's visibility, bumped from private to pub(super) so
letter.rs's parse_cover_letter can still call it, and one dangling
intra-doc-link comment fixed on LetterHead.initials).

Public surface identical for every consumer: typst_engine/mod.rs's
pub(crate) use re-exports (letterhead_initials, is_letterhead_name) now
point at the new module instead of letter, so docx/mod.rs's
crate::export::typst_engine::is_letterhead_name(...) /
::letterhead_initials(...) call sites are untouched.

letter.rs: 1497 -> 1048 LOC. letterhead.rs: 519 LOC (under both the 1400
hard cap and the 600 soft-watch threshold). Test count preserved exactly
(15 in letter.rs + 18 in letterhead.rs = the original 33).

Gates: cargo test --lib --all-features (4137 passed), cargo test --test
architecture (r8_no_oversized_modules green), clippy --all-targets
--all-features --locked -D warnings (clean), fmt (clean).
